fixed test errors: moved validators to email form
